+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Fix for duplicated client names and groups (issue #8)
 * New option for client command execution method (-ccem)
 * Worker pkg - worker.go, cce.go
+* Parallel command execution (issue #5)
 ```
 
 ### 0.2.5 (2014-03-16)